+++ src/stores/chains.ts
@@ -1,4 +1,5 @@
 import { InitialItemsState } from "repond";
+import { EventInstance } from "../types";
 
 // Chains for running events in order
 
@@ -8,6 +9,7 @@ function getDefaultState() {
     liveEventIds: [] as string[], // the chain of event nodettes , ["showSpeech_abcd", "wait_abcd"]. when it's empty, the chain'sRemoved
     // addTime?
     canAutoActivate: true, // whether events in the chain can be auto started, starts as false for addSubEvents, and can be set to true to activate
+    duplicateEventsToAdd: {} as Record<string, EventInstance>, // if an event is added, but the liveId already exists, it's added here, and that liveEvent will be canceled , once that event is removed, this event will be added to the end of the chain
   };
 }